Create a BPro integration proof-of-concept command
This Symfony console command is built to demonstrate the feasibility
of building an integration with BetterSoftware BPro inegration.

Since it is a code not mean to be run in production, some software
engineering practices are not done, e.g. documentation, code
commenting, etc. And the code is not built to for reuse with lots
of flexibility.

It is built using Symfony console command component only.